.Net 6 style
